django_event_mgnt_api/requirements.txt
Django
djangorestframework
django-filter
djangorestframework-simplejwt
python-dotenv
django-filter